requirements.txt
beautifulsoup4==4.12.0
discord.py==2.2.2
Flask==2.0.1
Flask_RESTful==0.3.9
psycopg2==2.9.3
pymongo==4.3.2
requests==2.28.1
roblox==2.0.0
APScheduler==3.10.4